Migrate feeds_lang().
