+++ view/tag/__init__.py
@@ -7,9 +7,9 @@ import sqlalchemy as sqla
 from model import get_session
 from model.orm.feeds import Feed, Tag
 from model.schema.feeds import Language
-from model.utils import all_langs_feeds
 from model.utils.custom import delete_feeds_tagged, insert_feeds_untagged
 
+from . import db
 from ..req import base_context
 
 bp = Blueprint("tag", __name__, url_prefix="/tag")
@@ -25,8 +25,8 @@ def tag():
             **base_context(),
             topnav_title=tag_row.Name,
             tag_row=tag_row,
-            feeds_lang=feeds_lang(tag_id),
-            feeds_lang_not=feeds_lang(tag_id, False),
+            feeds_lang=db.feeds_lang(tag_id),
+            feeds_lang_not=db.feeds_lang(tag_id, False),
     )
 
 @bp.route("/toggle_feeds", methods=['POST'])
@@ -41,25 +41,3 @@ def toggle_feeds():
 
     return ("", 200)
 
-def feeds_lang(tag_id, flag=True):
-    res = dict()
-
-    for lang_it in all_langs_feeds():
-        lang_it = Language(lang_it)
-
-        q = sqla.select(
-            Feed
-        ).where(
-            Feed.Language == lang_it.name,
-        ).order_by(
-            sqla.collate(Feed.Title, 'NOCASE')
-        )
-        if flag:
-            q = q.where(Feed.tags.any(Tag.TagID == tag_id))
-        else:
-            q = q.where(~Feed.tags.any(Tag.TagID == tag_id))
-
-        with get_session() as session:
-            res[lang_it] = [e[0] for e in session.execute(q)]
-
-    return res
